binding_analysis/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
scipy>=1.7.0
statsmodels>=0.12.0
sympy>=1.8
pytest>=6.0
pytest-cov>=2.12
pyyaml>=5.4.0
tabulate>=0.8.9
pathlib2>=2.3.6; python_version < '3.4'
